osprey: consolidate engine-AST helpers into _engine_ast_utils
Rename _features_ast_utils.py to _engine_ast_utils.py and move
collect_name_references (formerly _collect_name_references) from
features.py into it. Both features.py and the upcoming rules.py
import the three helpers from one place.
